Added a event card component to maintain sync
