+++ package.json
@@ -34,6 +34,6 @@
     "postcss": "^8.4.28",
     "react-router-dom": "^6.15.0",
     "tailwindcss": "^3.3.3",
-    "vite": "^6.2.3"
+    "vite": "^6.3.5"
   }
 }
